requirements.txt
albumentations==2.0.0
comet_ml==3.47.6
coremltools==8.1
GitPython==3.1.44
GitPython==3.1.44
hydra-core==1.3.2
ipython==8.12.3
matplotlib==3.10.0
mss==10.0.0
numpy==2.2.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
onnxsim==0.4.36
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
openvino==2024.6.0
paddle==1.0.2
pafy==0.5.5
pandas==2.2.3
Pillow==11.1.0
protobuf==5.29.3
psutil==6.0.0
pycocotools==2.0.8
pytesseract==0.3.13
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
seaborn==0.13.2
setuptools==75.6.0
streamlit==1.41.1
tensorflow==2.18.0
tensorflow_intel==2.18.0
tensorflowjs==4.22.0
tensorrt==10.7.0
tflite_runtime==2.14.0
tflite_support==0.4.4
thop==0.1.1.post2209072238
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
ultralytics_thop==2.0.13
x2paddle==1.6.0
